Add navigations tabs * create Views for Login, Sklep(Index), Koszyk, Zamowienia, Konto * create SklepController * add example of rendering loops from models on Index page, create example Towar class in Models
