+++ SklepInternetowy.WWW/Models/Towar.cs
@@ -1,0 +1,8 @@
+﻿namespace SklepInternetowy.WWW.Models
+{
+    public class Towar
+    {
+        public string? nazwa {  get; set; }
+        public decimal? cena { get; set; }
+    }
+}
