+++ SklepInternetowy.WWW/Views/Sklep/Konto.cshtml
@@ -1,0 +1,6 @@
+﻿@{
+    ViewData["Title"] = "Moje konto";
+}
+<h1>@ViewData["Title"]</h1>
+
+<p>Tu bedzie zakladka Konto
