+++ SklepInternetowy.WWW/Views/Sklep/Koszyk.cshtml
@@ -1,0 +1,6 @@
+﻿@{
+    ViewData["Title"] = "Koszyk";
+}
+<h1>@ViewData["Title"]</h1>
+
+<p>Tu bedzie wyswietlany koszyk</p>
