+++ SklepInternetowy.WWW/Views/Sklep/Login.cshtml
@@ -1,0 +1,6 @@
+﻿@{
+    ViewData["Title"] = "Login";
+}
+<h1>@ViewData["Title"]</h1>
+
+<p>Tu bedzie strona z logowaniem
