+++ SklepInternetowy.WWW/Views/Sklep/Zamowienia.cshtml
@@ -1,0 +1,6 @@
+﻿@{
+    ViewData["Title"] = "Moje zamowienia";
+}
+<h1>@ViewData["Title"]</h1>
+
+<p>Tu beda wyswietlane zamowienia
